Pure cosmetics
Moved three extractors into Extractor
