+++ algebra/src/main/scala/com/phasmidsoftware/number/algebra/eager/Extractors.scala
@@ -1,7 +1,12 @@
 package com.phasmidsoftware.number.algebra.eager
 
-import com.phasmidsoftware.number.algebra.core.{Q, Z}
-import com.phasmidsoftware.number.core.numerical.Fuzziness
+import com.phasmidsoftware.number.algebra.core.{CanPower, Q, Z}
+import com.phasmidsoftware.number.core.inner.Rational
+import com.phasmidsoftware.number.core.numerical
+import com.phasmidsoftware.number.core.numerical.Complex.convertToCartesian
+import com.phasmidsoftware.number.core.numerical.{ComplexCartesian, ComplexPolar, Fuzziness}
+
+import scala.annotation.tailrec
 
 /**
   * The Extractors object provides utility methods and patterns for extracting
@@ -83,6 +88,91 @@ object IsInteger {
       Some(z.toInt)
     case q: Q =>
       q.toRational.maybeInt
+    case _ =>
+      None
+  }
+}
+
+/**
+  * A utility object for pattern matching operations involving positive square roots.
+  * The main purpose of this object is to identify whether a given number or expression
+  * can be represented as the square root of another number.
+  */
+object IsSquareRoot {
+  /**
+    * Extractor method used to determine if the given `Eager` instance can be
+    * decomposed into a positive square root of a number.
+    *
+    * @param x the `Eager` instance to be analyzed
+    * @return an `Option` containing the square root as a `Number` if `x` represents a
+    *         positive square root of a number; `None` otherwise
+    */
+  def unapply(x: Eager): Option[Number] = x match {
+    case InversePower(2, n: Q) if n.signum > 0 =>
+      n.toRational.sqrt.toOption.map(RationalNumber(_))
+    case InversePower(2, r: Real) if r.signum > 0 =>
+      r power Rational.half
+    case InversePower(2, n: CanPower[Number] @unchecked) if n.signum > 0 =>
+      n.pow(RationalNumber.half)
+    case _ =>
+      None
+  }
+}
+
+/**
+  * An extractor object used for pattern matching to identify and extract `Eager` instances
+  * that satisfy specific conditions related to imaginary or irrational numbers.
+  */
+object IsImaginary {
+  /**
+    * Extracts the square root of the negative of the underlying Number of an InversePower if the number
+    * is negative and the power is 2.
+    *
+    * TODO: extend to match any purely imaginary Complex(0, x) where x != 0
+    * Currently only matches i = (-1)^(1/2)
+    * In this case, we would move this into Extractors.
+    *
+    * @param x the `Eager` instance to be matched and extracted.
+    * @return an `Option` containing the extracted `Number` if the input matches the expected pattern, or `None` otherwise.
+    */
+  def unapply(x: Eager): Option[Eager] = x.normalize match {
+    case InversePower(2, q: Q) if q.signum < 0 =>
+      q.toRational.negate.sqrt.toOption.map(RationalNumber(_).normalize)
+    case InversePower(2, r: Real) if r.signum < 0 =>
+      r.negate power Rational.half
+    case InversePower(2, n: CanPower[Number] @unchecked) if n.signum < 0 =>
+      for {
+        z: Real <- n.convert(Real.zero)
+        y <- z.negate.power(Rational.half)
+      } yield y.normalize
+    case x =>
+      None
+  }
+}
+
+
+/**
+  * The `HasImaginary` object provides utilities for working with complex mathematical objects,
+  * particularly for extracting and transforming specific patterns in `Eager` instances.
+  */
+object HasImaginary {
+  /**
+    * Extractor method for optionally extracting the imaginary part of an `Eager` instance.
+    * If the input is a complex number, it returns the imaginary part as an `Eager` instance.
+    * If the input is an InversePower with a negative value, then it returns the square root of the negated value.
+    *
+    * @param x the input value of type `Eager` to be analyzed and transformed.
+    * @return an `Option` containing a transformed `Eager` instance if a match is found,
+    *         or `None` if no pattern matches the input.
+    */
+  @tailrec
+  def unapply(x: Eager): Option[Eager] = x match {
+    case Complex(ComplexCartesian(_, i)) =>
+      Some(Eager(numerical.Real(i)))
+    case Complex(cp: ComplexPolar) =>
+      unapply(Complex(convertToCartesian(cp)))
+    case IsImaginary(z) =>
+      Some(z)
     case _ =>
       None
   }
