+++ algebra/src/main/scala/com/phasmidsoftware/number/algebra/eager/InversePower.scala
@@ -787,64 +787,3 @@ object InversePower {
     */
   implicit val showInversePower: Show[InversePower] = Show.show(_.render)
 }
-
-/**
-  * A utility object for pattern matching operations involving positive square roots.
-  * The main purpose of this object is to identify whether a given number or expression
-  * can be represented as the square root of another number.
-  *
-  * TODO move this to Extractors
-  */
-object IsSquareRoot {
-  /**
-    * Extractor method used to determine if the given `Eager` instance can be
-    * decomposed into a positive square root of a number.
-    *
-    * @param x the `Eager` instance to be analyzed
-    * @return an `Option` containing the square root as a `Number` if `x` represents a
-    *         positive square root of a number; `None` otherwise
-    */
-  def unapply(x: Eager): Option[Number] = x match {
-    case InversePower(2, n: Q) if n.signum > 0 =>
-      n.toRational.sqrt.toOption.map(RationalNumber(_))
-    case InversePower(2, r: Real) if r.signum > 0 =>
-      r power Rational.half
-    case InversePower(2, n: CanPower[Number] @unchecked) if n.signum > 0 =>
-      n.pow(RationalNumber.half)
-    case _ =>
-      None
-  }
-}
-
-/**
-  * An extractor object used for pattern matching to identify and extract `Number` instances
-  * that satisfy specific conditions related to imaginary or irrational numbers.
-  */
-object IsImaginary {
-  /**
-    * Extracts the square root of the negative of the underlying Number of an InversePower if the number
-    * is negative and the power is 2.
-    *
-    * TODO: extend to match any purely imaginary Complex(0, x) where x != 0
-    * Currently only matches i = (-1)^(1/2)
-    * In this case, we would move this into Extractors.
-    *
-    * TODO move this to Extractors
-    *
-    * @param x the `Eager` instance to be matched and extracted.
-    * @return an `Option` containing the extracted `Number` if the input matches the expected pattern, or `None` otherwise.
-    */
-  def unapply(x: Eager): Option[Eager] = x.normalize match {
-    case InversePower(2, q: Q) if q.signum < 0 =>
-      q.toRational.negate.sqrt.toOption.map(RationalNumber(_).normalize)
-    case InversePower(2, r: Real) if r.signum < 0 =>
-      r.negate power Rational.half
-    case InversePower(2, n: CanPower[Number] @unchecked) if n.signum < 0 =>
-      for {
-        z: Real <- n.convert(Real.zero)
-        y <- z.negate.power(Rational.half)
-      } yield y.normalize
-    case x =>
-      None
-  }
-}
